ground lives

+++ include/code_8002774.h
@@ -7,7 +7,7 @@ void sub_800290C(PixelPos *param_1, s32 param_2);
 void sub_8002934(PixelPos *param_1, PixelPos *param_2, PixelPos *param_3, s32 param_4, s32 param_5);
 s32 sub_8002A70(s32 a0, s32 a1, s32 a2);
 s32 sub_8002B04(s32, s32);
-s8 VecDirection8Radial(PixelPos *param_1);
+s32 VecDirection8Radial(PixelPos *param_1);
 s8 SizedDeltaDirection4(const PixelPos *r0, const PixelPos *r1, const PixelPos *r2, const PixelPos *r3);
 s8 SizedDeltaDirection8(const PixelPos *r0, const PixelPos *r1, const PixelPos *r2, const PixelPos *r3);
